+++ src/resolvers.js
@@ -64,7 +64,7 @@ const enumeratedProperties = schema => {
 const arraysOfStrings = schema => {
   if (schema.type === 'array' && schema.items && !(Array.isArray(schema.items)) && ['string', 'number', 'integer'].includes(schema.items.type)) {
     if (schema.format === 'choices') return 'arrayChoices'
-    if (schema.uniqueItems) {
+    if (schema.uniqueItems && schema.items.enum) {
       /* if 'selectize' enabled it is expected to be selectized control */
       if (schema.format === 'selectize') return 'arraySelectize'
       if (schema.format === 'select2') return 'arraySelect2'