+++ src/editors/object.js
@@ -1041,7 +1041,9 @@ export class ObjectEditor extends AbstractEditor {
   }
 
   onOutsideModalClick (e) {
-    if (this.addproperty_holder && !this.addproperty_holder.contains(e.target) && this.adding_property) {
+    if (this.addproperty_holder &&
+      !this.addproperty_holder.contains(e.path[0] || e.composedPath()[0]) &&
+      this.adding_property) {
       e.preventDefault()
       e.stopPropagation()
       this.toggleAddProperty()