+++ src/iconlibs/index.js
@@ -1,5 +1,5 @@
 // import  { bootstrap2Iconlib } from  './bootstrap2.js'
-// import  { bootstrap3Iconlib } from  './bootstrap3.js'
+import { bootstrap3Iconlib } from './bootstrap3.js'
 import { fontawesome3Iconlib } from './fontawesome3.js'
 import { fontawesome4Iconlib } from './fontawesome4.js'
 import { fontawesome5Iconlib } from './fontawesome5.js'
@@ -11,7 +11,7 @@ import { spectreIconlib } from './spectre.js'
 
 export const iconlibs = {
   // bootstrap2: bootstrap2Iconlib,
-  // bootstrap3: bootstrap3Iconlib,
+  bootstrap3: bootstrap3Iconlib,
   fontawesome3: fontawesome3Iconlib,
   fontawesome4: fontawesome4Iconlib,
   fontawesome5: fontawesome5Iconlib,