+++ tests/codeceptjs/editors/object_test.js
@@ -145,3 +145,15 @@ Scenario('should have unique ids', (I) => {
   I.click('.form-group:nth-child(2) .form-check-input')
   I.waitForText('i am actually a cat')
 });
+
+Scenario('should hide properties with unfulfilled dependencies', (I) => {
+  I.amOnPage('object-with-dependencies.html');
+  I.seeElement('[data-schemapath="root.enable_option"] input');
+  I.dontSeeElement('[data-schemapath="root.make_new"] input');
+  I.dontSeeElement('[data-schemapath="root.existing_name"] input');
+
+  I.click('[data-schemapath="root.enable_option"] input');
+  I.seeElement('[data-schemapath="root.enable_option"] input');
+  I.seeElement('[data-schemapath="root.make_new"] input');
+  I.dontSeeElement('[data-schemapath="root.existing_name"] input');
+});