+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - AbstractEditor, AbstractTheme, AbstractIconlib had been removed, re-expose them
 - fix static property #724 #723
+- fix option remove_empty_properties #725 #728
 
 ### 2.2.0