+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 2.3.0-dev
 
 - Removed codeception container, use `codeceptjs` as node-dev module
+- Enabled Bootstrap 3 theme/icons
 
 ### 2.2.1-current
 
@@ -53,6 +54,7 @@
   - Fix of #629 - BS4 theme bug where basic tab was placed outside tabs container.
   - Improved IMask support for `Date`, `Number`, `IMask.MaskedEnum`, `IMask.MaskedRange` and regular expression masks. #591
   - Added support for recursive callback options.
+  - Fix of #692 - Resolves an issue where modal click detection was not working when the editor is attached inside a shadow DOM by changing the event target to use the path.
 
 ### 2.0.0-alpha-1