+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Unreleased
 - added option `max_depth` used to specify the maximum depth of level's schema that have to be rendered
 - added option `use_default_values` used to specify if default values based on the "type" of the property have to be used
+- when `use_default_values` is false, number and integer fields have undefined value when input is empty
+- when `use_default_values` is false, string fields have undefined value at the beginning. After that user edit the field, empty input is threated as an empty string
 
 ### 2.3.0-dev